impl most of host-side plugin interface; refactor
