+++ .idea/assemblylift.iml
@@ -14,6 +14,8 @@
       <sourceFolder url="file://$MODULE_DIR$/backends/aws-lambda/host/src" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/examples/todo/create/src" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/cli/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/backends/aws-lambda/iomod/plugins/dynamodb/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/backends/aws-lambda/iomod/plugins/dynamodb/guest/src" isTestSource="false" />
       <excludeFolder url="file://$MODULE_DIR$/target" />
     </content>
     <orderEntry type="inheritedJdk" />
