+++ Cargo.toml
@@ -4,6 +4,8 @@ members = [
     "backends/aws-lambda/guest",
     "backends/aws-lambda/iomod",
     "backends/aws-lambda/iomod/guest",
+    "backends/aws-lambda/iomod/plugins/dynamodb",
+    "backends/aws-lambda/iomod/plugins/dynamodb/guest",
     "core",
     "core/guest",
     "core/event",
