+++ core/src/iomod/mod.rs
@@ -7,14 +7,19 @@ use std::io::ErrorKind;
 use std::sync::Mutex;
 
 use crossbeam_utils::atomic::AtomicCell;
-
-use wasmer_runtime::memory::MemoryView;
 use wasmer_runtime::Ctx;
+use wasmer_runtime::memory::MemoryView;
 use wasmer_runtime_core::vm;
 
-use crate::WasmBufferPtr;
 use assemblylift_core_event::threader::Threader;
 use assemblylift_core_event_common::constants::EVENT_BUFFER_SIZE_BYTES;
+
+use crate::iomod::registry::ModuleRegistry;
+use crate::WasmBufferPtr;
+
+pub mod macros;
+pub mod registry;
+pub mod plugin;
 
 lazy_static! {
     pub static ref MODULE_REGISTRY: Mutex<ModuleRegistry> = Mutex::new(ModuleRegistry::new());
@@ -25,23 +30,7 @@ fn to_io_error<E: Error>(err: E) -> io::Error {
 }
 
 pub trait IoModule {
-    fn register(registry: &mut ModuleRegistry); // MAYBE
-}
-
-pub type AsmlAbiFn = fn(&mut vm::Ctx, WasmBufferPtr, WasmBufferPtr, u32) -> i32;
-pub type ModuleMap = HashMap<String, HashMap<String, HashMap<String, AsmlAbiFn>>>;
-
-#[derive(Clone)]
-pub struct ModuleRegistry {
-    pub modules: ModuleMap,
-}
-
-impl ModuleRegistry {
-    pub fn new() -> Self {
-        ModuleRegistry {
-            modules: Default::default(),
-        }
-    }
+    fn register(registry: &mut ModuleRegistry);
 }
 
 pub fn asml_abi_invoke(
@@ -154,69 +143,3 @@ pub fn spawn_event(
     event_id as i32
 }
 
-#[macro_export]
-macro_rules! register_calls {
-    ($reg:expr, $org_name:ident => { $ns_name:ident => $ns:tt })
-    => {{
-        let org_name = stringify!($org_name);
-        let ns_name = stringify!($ns_name);
-
-        let mut namespace_map = HashMap::new();
-
-        // $({
-            let mut name_map = __register_calls!($ns);
-            namespace_map.entry(ns_name.to_string()).or_insert(name_map);
-        // })*
-
-        $reg.modules.entry(org_name.to_string()).or_insert(namespace_map);
-    }};
-}
-
-#[macro_export]
-#[doc(hidden)]
-macro_rules! __register_calls {
-    ({ $( $call_name:ident => $call:expr ),* $(,)? }) => {{
-        let mut name_map = HashMap::new();
-        $(
-            let call_name = String::from(stringify!($call_name));
-            name_map.entry(call_name).or_insert($call as AsmlAbiFn);
-        )*
-        name_map
-    }};
-}
-
-#[macro_export]
-macro_rules! call {
-    ($call_name:ident => $call:item) => {
-        $call
-
-        pub fn $call_name (ctx: &mut vm::Ctx, mem: WasmBufferPtr, input: WasmBufferPtr, input_len: u32) -> i32 {
-            use assemblylift_core::iomod::spawn_event;
-
-            println!("TRACE: {}", stringify!($call_name));
-            let input_vec = __wasm_buffer_as_vec!(ctx, input, input_len);
-            let call = paste::expr! { [<$call_name _impl>] }(input_vec);
-            spawn_event(ctx, mem, call)
-        }
-    };
-}
-
-#[macro_export]
-#[doc(hidden)]
-macro_rules! __wasm_buffer_as_vec {
-    ($ctx:ident, $input:ident, $input_len:ident) => {{
-        let wasm_instance_memory = $ctx.memory(0);
-        let input_deref: &[AtomicCell<u8>] = match $input.deref(wasm_instance_memory, 0, $input_len)
-        {
-            Some(memory) => memory,
-            None => panic!("could not dereference WASM guest memory in __wasm_buffer_as_vec"),
-        };
-
-        let mut as_vec: Vec<u8> = Vec::new();
-        for (idx, b) in input_deref.iter().enumerate() {
-            as_vec.insert(idx, b.load());
-        }
-
-        as_vec
-    }};
-}
